Resolve copilot review comments
