+++ tests/Benchmark/Benchmark.cpp
@@ -20,6 +20,10 @@
 #include "Benchmark.h"
 #include <qa_interfaces/json/JBenchmark.h>
 
+#include <cinttypes>
+#include <cmath>
+#include <cstdio>
+
 namespace Thunder {
 namespace BenchmarkMemory {
     Exchange::IMemory* MemoryObserver(const RPC::IRemoteConnection* connection)
@@ -86,6 +90,92 @@ namespace Plugin {
         );
     }
 
+    namespace {
+
+        struct LatencyAccumulator {
+            std::vector<uint64_t> samplesUs;
+
+            void Record(uint64_t us)
+            {
+                samplesUs.push_back(us);
+            }
+
+            QualityAssurance::IBenchmark::RoundTripStats Stats() const
+            {
+                QualityAssurance::IBenchmark::RoundTripStats stats{};
+                if (samplesUs.empty()) return stats;
+
+                uint64_t sum = 0;
+                uint64_t minVal = UINT64_MAX;
+                uint64_t maxVal = 0;
+
+                for (auto v : samplesUs) {
+                    sum += v;
+                    if (v < minVal) minVal = v;
+                    if (v > maxVal) maxVal = v;
+                }
+
+                uint64_t avg = sum / samplesUs.size();
+
+                double variance = 0.0;
+                for (auto v : samplesUs) {
+                    double diff = static_cast<double>(v) - static_cast<double>(avg);
+                    variance += diff * diff;
+                }
+                variance /= samplesUs.size();
+
+                // Convert µs to ns for the interface
+                stats.minNs = minVal * 1000;
+                stats.avgNs = avg * 1000;
+                stats.maxNs = maxVal * 1000;
+                stats.stddevNs = static_cast<uint64_t>(std::sqrt(variance)) * 1000;
+
+                return stats;
+            }
+        };
+
+        template<typename CALLABLE>
+        QualityAssurance::IBenchmark::BenchmarkResult MeasurePayloadMethod(
+            const string& apiName, uint32_t iterations, CALLABLE&& fn)
+        {
+            Core::ProcessInfo processInfo(Core::ProcessInfo().Id());
+
+            uint64_t residentBefore = processInfo.Resident();
+            uint64_t allocatedBefore = processInfo.Allocated();
+
+            LatencyAccumulator acc;
+
+            // Single pass: time each call individually, derive avg from the sum
+            Core::StopWatch batchTimer;
+            for (uint32_t i = 0; i < iterations; i++) {
+                Core::StopWatch timer;
+                fn();
+                uint64_t elapsedUs = timer.Elapsed();
+                acc.Record(elapsedUs);
+            }
+            uint64_t totalBatchUs = batchTimer.Elapsed();
+
+            uint64_t residentAfter = processInfo.Resident();
+            uint64_t allocatedAfter = processInfo.Allocated();
+
+            QualityAssurance::IBenchmark::BenchmarkResult result{};
+            result.apiName = apiName;
+            result.iterations = iterations;
+            result.roundTrip = acc.Stats();
+            // Override avgNs with batch-derived value for better precision
+            if (iterations > 0) {
+                result.roundTrip.avgNs = (totalBatchUs * 1000) / iterations;
+            }
+            result.memory.residentBefore = residentBefore;
+            result.memory.residentAfter = residentAfter;
+            result.memory.allocatedBefore = allocatedBefore;
+            result.memory.allocatedAfter = allocatedAfter;
+
+            return result;
+        }
+
+    } // anonymous namespace
+
     const string Benchmark::Initialize(PluginHost::IShell* service)
     {
         ASSERT(_benchmark == nullptr);
@@ -105,25 +195,32 @@ namespace Plugin {
         if (_benchmark == nullptr) {
             result = _T("Couldn't create Benchmark instance");
         } else {
-            QualityAssurance::JBenchmark::Register(*this, _benchmark);
-
-            _benchmark->Register(&_benchmarkNotification);
-
-            if (_connectionId == 0) {
-                _memory = Thunder::BenchmarkMemory::MemoryObserver(nullptr);
-            } else {
-                const RPC::IRemoteConnection* connection = _service->RemoteConnection(_connectionId);
-                if (connection == nullptr) {
-                    result = _T("Benchmark crashed at initialize!");
-                } else {
-                    _memory = Thunder::BenchmarkMemory::MemoryObserver(connection);
-                    ASSERT(_memory != nullptr);
-                    connection->Release();
-                }
+            _payloadProxy = _benchmark->QueryInterface<QualityAssurance::IBenchmarkPayload>();
+            if (_payloadProxy == nullptr) {
+                result = _T("Couldn't obtain IBenchmarkPayload interface from Benchmark implementation");
             }
 
-            if (_memory == nullptr) {
-                result = _T("Benchmark could not instantiate a Memory observer!");
+            if (result.empty()) {
+                RegisterJsonRpcHandlers();
+
+                _benchmark->Register(&_benchmarkNotification);
+
+                if (_connectionId == 0) {
+                    _memory = Thunder::BenchmarkMemory::MemoryObserver(nullptr);
+                } else {
+                    const RPC::IRemoteConnection* connection = _service->RemoteConnection(_connectionId);
+                    if (connection == nullptr) {
+                        result = _T("Benchmark crashed at initialize!");
+                    } else {
+                        _memory = Thunder::BenchmarkMemory::MemoryObserver(connection);
+                        ASSERT(_memory != nullptr);
+                        connection->Release();
+                    }
+                }
+
+                if (_memory == nullptr && result.empty()) {
+                    result = _T("Benchmark could not instantiate a Memory observer!");
+                }
             }
         }
 
@@ -144,7 +241,12 @@ namespace Plugin {
         if (_benchmark != nullptr) {
             _benchmark->Unregister(&_benchmarkNotification);
 
-            QualityAssurance::JBenchmark::Unregister(*this);
+            UnregisterJsonRpcHandlers();
+
+            if (_payloadProxy != nullptr) {
+                _payloadProxy->Release();
+                _payloadProxy = nullptr;
+            }
 
             RPC::IRemoteConnection* connection(_service->RemoteConnection(_connectionId));
 
@@ -181,6 +283,195 @@ namespace Plugin {
     void Benchmark::BenchmarkCompleted()
     {
         QualityAssurance::JBenchmark::Event::PerformanceCheckCompleted(*this);
+    }
+
+    void Benchmark::RegisterJsonRpcHandlers()
+    {
+        PluginHost::JSONRPC::RegisterVersion(_T("JBenchmark"), 1, 0, 0);
+
+        // Method: 'trigger' — runs benchmarks via COM-RPC proxy
+        PluginHost::JSONRPC::Register<JsonData::Benchmark::TriggerParamsData, Core::JSON::Boolean>(_T("trigger"),
+            [this](const JsonData::Benchmark::TriggerParamsData& params, Core::JSON::Boolean& success) -> uint32_t {
+                if ((params.IsSet() == false) || (params.IsDataValid() == false)) {
+                    return Core::ERROR_BAD_REQUEST;
+                }
+
+                _adminLock.Lock();
+                _results.clear();
+                _adminLock.Unlock();
+
+                RunPayloadBenchmarks(params.Iterations);
+                ApplyThresholds();
+
+                QualityAssurance::JBenchmark::Event::PerformanceCheckCompleted(*this);
+
+                _adminLock.Lock();
+                bool allPassed = std::all_of(_results.begin(), _results.end(),
+                    [](const QualityAssurance::IBenchmark::BenchmarkResult& r) { return r.passed; });
+                _adminLock.Unlock();
+
+                success = allPassed;
+                TRACE(Trace::Information, (_T("Benchmark completed: %u iterations via COM-RPC proxy"),
+                    static_cast<uint32_t>(params.Iterations)));
+                return Core::ERROR_NONE;
+            });
+
+        // Method: 'collectdata' — returns shell-side results
+        PluginHost::JSONRPC::Register<void, Core::JSON::ArrayType<JsonData::Benchmark::BenchmarkResultData>>(_T("collectdata"),
+            [this](Core::JSON::ArrayType<JsonData::Benchmark::BenchmarkResultData>& report) -> uint32_t {
+                _adminLock.Lock();
+                report.Set(true);
+                for (const auto& r : _results) {
+                    report.Add() = r;
+                }
+                _adminLock.Unlock();
+                return Core::ERROR_NONE;
+            });
+
+        // Method: 'setthreshold' — stores thresholds on the shell side
+        PluginHost::JSONRPC::Register<JsonData::Benchmark::SetThresholdParamsData, Core::JSON::Boolean>(_T("setthreshold"),
+            [this](const JsonData::Benchmark::SetThresholdParamsData& params, Core::JSON::Boolean& success) -> uint32_t {
+                if ((params.IsSet() == false) || (params.IsDataValid() == false)) {
+                    return Core::ERROR_BAD_REQUEST;
+                }
+
+                _adminLock.Lock();
+                _maxLatencyDeviationPct = params.MaxLatencyDeviationPct;
+                _maxMemoryGrowthBytes = params.MaxMemoryGrowthBytes;
+                _baselines.clear();
+                _adminLock.Unlock();
+
+                success = true;
+                TRACE(Trace::Information, (_T("Thresholds set: latency=%.1f%%, memory=%" PRIu64 " bytes"),
+                    static_cast<float>(params.MaxLatencyDeviationPct),
+                    static_cast<uint64_t>(params.MaxMemoryGrowthBytes)));
+                return Core::ERROR_NONE;
+            });
+    }
+
+    void Benchmark::UnregisterJsonRpcHandlers()
+    {
+        PluginHost::JSONRPC::Unregister(_T("trigger"));
+        PluginHost::JSONRPC::Unregister(_T("collectdata"));
+        PluginHost::JSONRPC::Unregister(_T("setthreshold"));
+    }
+
+    void Benchmark::RunPayloadBenchmarks(uint32_t iterations)
+    {
+        ASSERT(_payloadProxy != nullptr);
+
+        auto addResult = [this](QualityAssurance::IBenchmark::BenchmarkResult&& r) {
+            _adminLock.Lock();
+            _results.push_back(std::move(r));
+            _adminLock.Unlock();
+        };
+
+        addResult(MeasurePayloadMethod("SendUint32", iterations, [this]() {
+            _payloadProxy->SendUint32(42);
+        }));
+
+        addResult(MeasurePayloadMethod("SendUint64", iterations, [this]() {
+            _payloadProxy->SendUint64(UINT64_C(123456789));
+        }));
+
+        addResult(MeasurePayloadMethod("SendString", iterations, [this]() {
+            _payloadProxy->SendString(_T("benchmark_payload_string"));
+        }));
+
+        addResult(MeasurePayloadMethod("SendSampleData", iterations, [this]() {
+            QualityAssurance::SampleData data;
+            data.id = 1;
+            data.value = 100;
+            data.name = _T("sample");
+            _payloadProxy->SendSampleData(data);
+        }));
+
+        addResult(MeasurePayloadMethod("SendWithNoParameters", iterations, [this]() {
+            _payloadProxy->SendWithNoParameters();
+        }));
+
+        addResult(MeasurePayloadMethod("SendReceiveUint32", iterations, [this]() {
+            uint32_t out = 0;
+            _payloadProxy->SendReceiveUint32(42, out);
+        }));
+
+        addResult(MeasurePayloadMethod("SendReceiveString", iterations, [this]() {
+            string out;
+            _payloadProxy->SendReceiveString(_T("benchmark_roundtrip"), out);
+        }));
+
+        addResult(MeasurePayloadMethod("SendReceiveSampleData", iterations, [this]() {
+            QualityAssurance::SampleData in;
+            in.id = 1;
+            in.value = 200;
+            in.name = _T("roundtrip");
+            QualityAssurance::SampleData out;
+            _payloadProxy->SendReceiveSampleData(in, out);
+        }));
+
+        addResult(MeasurePayloadMethod("Add", iterations, [this]() {
+            uint32_t result = 0;
+            _payloadProxy->Add(17, 25, result);
+        }));
+    }
+
+    void Benchmark::ApplyThresholds()
+    {
+        _adminLock.Lock();
+        bool hasThresholds = (_maxLatencyDeviationPct > 0.0f || _maxMemoryGrowthBytes > 0);
+
+        if (hasThresholds && _baselines.empty()) {
+            // First run with thresholds: store as baseline
+            for (auto& r : _results) {
+                _baselines[r.apiName] = r;
+                r.passed = true;
+            }
+        } else if (hasThresholds) {
+            // Subsequent runs: compare against baseline
+            for (auto& r : _results) {
+                r.passed = true;
+                r.failureReason = string();
+
+                auto baseline = _baselines.find(r.apiName);
+                if (baseline != _baselines.end()) {
+                    // Latency check
+                    if (_maxLatencyDeviationPct > 0.0f && baseline->second.roundTrip.avgNs > 0) {
+                        double baselineAvg = static_cast<double>(baseline->second.roundTrip.avgNs);
+                        double currentAvg = static_cast<double>(r.roundTrip.avgNs);
+                        double deviationPct = ((currentAvg - baselineAvg) / baselineAvg) * 100.0;
+                        if (deviationPct > static_cast<double>(_maxLatencyDeviationPct)) {
+                            r.passed = false;
+                            char buf[256];
+                            snprintf(buf, sizeof(buf), "latency deviation %.1f%% exceeds %.1f%% threshold (baseline=%" PRIu64 " ns, current=%" PRIu64 " ns)",
+                                deviationPct, static_cast<double>(_maxLatencyDeviationPct),
+                                baseline->second.roundTrip.avgNs, r.roundTrip.avgNs);
+                            r.failureReason = string(buf);
+                        }
+                    }
+                    // Memory check
+                    if (_maxMemoryGrowthBytes > 0) {
+                        int64_t growth = static_cast<int64_t>(r.memory.residentAfter) - static_cast<int64_t>(r.memory.residentBefore);
+                        if (growth > 0 && static_cast<uint64_t>(growth) > _maxMemoryGrowthBytes) {
+                            r.passed = false;
+                            char buf[256];
+                            snprintf(buf, sizeof(buf), "memory growth %" PRId64 " bytes exceeds %" PRIu64 " byte threshold",
+                                growth, _maxMemoryGrowthBytes);
+                            string memMsg(buf);
+                            if (!r.failureReason.empty()) {
+                                r.failureReason += "; ";
+                            }
+                            r.failureReason += memMsg;
+                        }
+                    }
+                }
+            }
+        } else {
+            // No thresholds: everything passes
+            for (auto& r : _results) {
+                r.passed = true;
+            }
+        }
+        _adminLock.Unlock();
     }
 
 }
